requirements.txt
flask
tensorflow-cpu==2.19.0
pillow
numpy
gunicorn
firebase-admin
cloudinary
python-dotenv